refactor: extract PendingApprovalCoordinator into AppState extension (Stage 4b)
Final stage of the AppState god-object split. Extracts the approve/deny
lifecycle, the alert-chain state machine (build 55-60 polish sprint),
and the lock-screen action static handler. After this lands, AppState.swift
hits the BACKLOG-sketched slim target.

Moves out of Clave/AppState.swift (12 methods + nested enum):
- dismissActiveAlert, dismissAllActiveAlerts
- refreshPendingRequests, purgeStalePendingRequests
- setKindOverride
- approvePendingRequest, denyPendingRequest, advanceChainPosition (private)
- static handlePendingApprovalAction (lock-screen action handler)
- static performApprove, performDeny, performRemovePending (private)
- enum ApproveOutcome (nested type)

Into the new file:
- Clave/AppState+PendingApprovalCoordinator.swift (276 LOC) — extension AppState

What stays in AppState.swift:
- Stored properties: pendingRequests, dismissedAlertRequestIds,
  processedInChain (Swift constraint)
- Computed properties: freshPendingRequests, activeApprovalRequest,
  pendingApprovalQueueDepth, chainPosition, chainTotal — kept on the
  class defensively (Stage 3b precedent for @Observable-bound computeds
  that drive UI binding)
- static let pendingRequestTTLSeconds — read by both class-side computed
  freshPendingRequests and extension-side purgeStalePendingRequests; static
  members visible across class+extension boundary so location is just
  about adjacency to consumers

Two AppState.swift access widenings (same pattern as bunkerSecretsTick
in Stage 4a):
- private var dismissedAlertRequestIds -> var (extension dismissActive*
  methods need to mutate cross-file)
- private(set) var processedInChain -> var (extension dismissAll*,
  refreshPendingRequests, advanceChainPosition need to mutate)

Zero behavior change:
- Function bodies preserved byte-for-byte
- private modifiers preserved on advanceChainPosition,
  performRemovePending (only called within extension file)
- Static methods relocate; AppState.handlePendingApprovalAction call
  sites in AppDelegate / NSE-side handler resolve identically since
  extension static methods are accessible via Type.method()
- Cross-extension calls (approvePendingRequest -> Self.performApprove,
  denyPendingRequest -> Self.performDeny) all stay in same extension file

Test impact: zero changes.
- AppStatePendingApprovalTests calls appState.<method> via extension
  dispatch + AppState.<staticMethod> via type dispatch — works unchanged
- LockScreenActionRoutingTests calls AppState.handlePendingApprovalAction
  via static dispatch — works unchanged

External callers unchanged:
- MainTabView (refresh / purge on scenePhase)
- InboxView (approve / deny / dismiss)
- PendingRequestDetailView (approve / deny / setKindOverride)
- HomeView (dismiss handlers, root alert binding)
- AppDelegate (lock-screen action handler -> static
  handlePendingApprovalAction)

No pbxproj edits — Clave/ directory is auto-synced.

Verification:
- xcodebuild test -skip-testing:ClaveUITests on iPhone 17 / iOS 26.4:
  - Pre-baseline (main @ 3b35719): 231 passed / 0 failed
  - Post-extraction (this branch):  231 passed / 0 failed
  - ** TEST SUCCEEDED ** in both runs

AppState.swift: 611 -> 348 LOC (-263, -43.0%).
SPRINT TOTAL: 2,338 -> 348 LOC (-85.1% from sprint start).

This completes the AppState god-object refactor sprint:
- Stage 1: RelayUtils.swift (78 LOC pure utilities)
- Stage 2: legacy migration deletion (-524 LOC dead code)
- Stage 3a: AppState+NostrConnect.swift (NIP-46 handshake)
- Stage 3b: AppState+ProfileFetcher.swift (kind:0 fetch)
- Stage 3c: AppState+ProxyClient.swift (NIP-98 + retry queue)
- Petname removal (-108 LOC dead code)
- Stage 4a: AppState+AccountManager.swift (account lifecycle)
- Stage 4b: AppState+PendingApprovalCoordinator.swift (alert chain) - this PR

AppState.swift is now a slim @Observable container with init,
loadState coordinator, and the @Observable storage + computed
property surface.
